+++ js/data.js
@@ -1,0 +1,32 @@
+const COMMETS = ['Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'];
+
+const NAMES = ['Иван', 'Пётр', 'Анастасия', 'Николай', 'Мария', 'Егор', 'Александр', 'Екатерина'];
+
+const ID = [];
+const ID_MAX = 25;
+const ID_MIN = 1;
+const LIKE_MIN = 15;
+const LIKE_MAX = 200;
+const COMMENT_MIN = 0;
+const COMMENT_MAX = 30;
+const AVATAR_MIN = 1;
+const AVATAR_MAX = 6;
+
+export {
+  COMMETS,
+  NAMES,
+  ID,
+  ID_MAX,
+  ID_MIN,
+  LIKE_MIN,
+  LIKE_MAX,
+  COMMENT_MIN,
+  COMMENT_MAX,
+  AVATAR_MIN,
+  AVATAR_MAX,
+};
